==> tutorial/2_fuzhi.ipynb <==
import os              
os.environ['PYTHONHASHSEED'] = '0'
import desc          
import pandas as pd                                                    
import numpy as np                                                     
import scanpy as sc                                                                                 
from time import time                                                       
import sys
import matplotlib
import matplotlib.pyplot as plt

import tensorflow as tf

import anndata as ad

adata = sc.read_h5ad('/home/huggs/shiyi/SCST/RCTD/RCTD/predata/celltrek_kidney/celltrek_kidney/Mouse_Adult_DGE.h5ad')

adata

adata.obs

adata.var

print(adata.raw)

adata_train=sc.read_h5ad("/home/huggs/shiyi/SCST/RCTD/RCTD/predata/celltrek_kidney/mymodel/2_desc/2_desc_222copy/sc_desc.h5ad")

adata_train

adata_train.obs

adata_train.var

print(adata.raw.X)

print(adata_train.raw.X)

################新变量###############
#adata.raw.X赋给adata_train.X
import scipy.sparse
my_matrix=adata.raw.X
type(my_matrix)
my_array=my_matrix.A
my_array

#adata_train.X换成adata.raw.X,adata_train.var换成adata.var
adata_trainnew = ad.AnnData(my_array, obs=adata.obs, var=adata.var)
adata_trainnew.uns=adata_train.uns
adata_trainnew.obsm=adata_train.obsm
adata_trainnew.obsp=adata_train.obsp
adata_trainnew.obs
adata_trainnew

type(adata_trainnew.X)

print(adata_trainnew.raw)

print(adata.raw)

adata.raw.var.index

adata_trainnew.raw

adata_trainnew.var.index

#INDEX
adata.raw.var.index = adata_trainnew.var.index
adata.raw.var

adata.__dict__['_raw'].__dict__['_var'] = adata.__dict__['_raw'].__dict__['_var'].rename(columns={'_index': 'features'})
adata.raw.var

#adata.raw赋给adata_trainnew
adata_trainnew.raw = adata
adata_trainnew.raw

print(adata_trainnew.raw)

adata_trainnew.raw.var

adata_trainnew

adata_trainnew.obs

adata_trainnew.obs['desc_0.8']=adata_train.obs['desc_0.8']

adata_trainnew.var_names

#adata_trainnew.obs_names=adata.obs_names

#adata_trainnew.write_h5ad('/home/huggs/shiyi/SCST/RCTD/RCTD/c2l_brain/test/sc_desc_new.h5ad')
adata_trainnew.write_h5ad('/home/huggs/shiyi/SCST/RCTD/RCTD/predata/celltrek_kidney/mymodel/6_DE/sc_desc_anno.h5ad')

adata_trainnew=sc.read_h5ad('/home/huggs/shiyi/SCST/RCTD/RCTD/predata/celltrek_kidney/mymodel/6_DE/sc_desc_anno.h5ad')

adata_trainnew

adata_trainnew.obs

adata = sc.read_h5ad('/home/huggs/shiyi/SCST/RCTD/RCTD/predata/celltrek_kidney/celltrek_kidney/Mouse_Adult_DGE.h5ad')
adata

adata_trainnew.X

print(adata_trainnew.raw.X)

type(adata_train.X)

type(adata_trainnew.X)
