==> tutorial/4_inputdata.ipynb <==
import sys
import scanpy as sc
import anndata
import pandas as pd
import numpy as np
import os
from scipy.io import mmwrite
import desc
import scipy
import anndata as ad

test=sc.read_h5ad("testdata_undersample.h5ad")
test

print(test.X)

train=sc.read_h5ad("traindata_undersample.h5ad")
train

print(train.X)

valid=sc.read_h5ad("validation_undersample.h5ad")
valid

print(valid.X)

desc.normalize_per_cell(train, counts_per_cell_after=1e4)
sc.pp.log1p(train)
desc.scale(train, zero_center=True, max_value=3)
train

desc.normalize_per_cell(valid, counts_per_cell_after=1e4)
sc.pp.log1p(valid)
desc.scale(valid, zero_center=True, max_value=3)
valid

desc.normalize_per_cell(test, counts_per_cell_after=1e4)
sc.pp.log1p(test)
desc.scale(test, zero_center=True, max_value=3)
test    #non norm: 80976 × 2977

train.obs['data']='traindata'
test.obs['data']='testdata'
valid.obs['data']='validation'

test.write_h5ad("test_undersample_norm.h5ad")

test[test.obs['batch']=='st'].obs_names

##合并
import anndata as ad
adatas=[train,test,valid]
adatas=ad.concat(adatas,merge='same')
adatas

#combine try 2
import anndata as ad
combined = ad.concat([train,test,valid], join="outer")
combined

combined.obs['desc_0.8']=combined.obs['desc_0.8'].astype('category')

combined.obs['desc_0.8']

combined.obs['batch']

combined.write_h5ad("inputdata_celltrek_kidney.h5ad")

combined.obs['data'].value_counts()

combined.X
