==> tutorial/4_train_valid.ipynb <==
import sys
import scanpy as sc
import anndata
import pandas as pd
import numpy as np
import os
from scipy.io import mmwrite
import desc
import scipy
import anndata as ad

###merge sc_combine&st_combine###

## sc

#desc_0.8
sc_combine=sc.read_h5ad("/home/huggs/shiyi/SCST/RCTD/RCTD/predata/celltrek_kidney/mymodel/2_desc/2_desc_222copy/sc_desc.h5ad")
sc_combine

print(sc_combine.X)

print(sc_combine.raw.X)

sc_combine.X=sc_combine.raw.X

#add batch
sc_combine.obs["batch"]="sc"

#add annotation_1: desc_0.2
#adata_snrna_raw=sc.read_h5ad("/s/f/shiyi/ST/Data/cell2location/data/filter_raw.h5ad")


len(sc_combine.obs["desc_0.8"].cat.categories)

sc_combine.obs

## st

st_combine=sc.read_h5ad("/home/huggs/shiyi/SCST/RCTD/RCTD/predata/celltrek_kidney/mymodel/1_predata/st_combine.h5ad")
st_combine

print(st_combine.X)

st_combine.obs["batch"]="st"

st_combine.obs

st_rctd=sc.read_h5ad("/home/huggs/shiyi/SCST/RCTD/RCTD/predata/celltrek_kidney/mymodel/3_RCTD/st_combine_subset.h5ad")
st_rctd

print(st_rctd.X)

st_rctd.obs["batch"]="st"

## merge

#合并ST,adata
import anndata as ad
adatas=[sc_combine,st_rctd]
adatas=ad.concat(adatas,merge='same')
adatas

#combine try 2
import anndata as ad
combined = ad.concat([sc_combine,st_rctd], join="outer")
combined

combined.var=adatas.var

combined.obs['desc_0.8']

#保存报错TypeError: Can't implicitly convert non-string objects to strings  Above error raised while writing key 'desc_0.4' of <class 'h5py._hl.group.Group'> from /.  Above error raised while writing key 'obs' of <class 'h5py._hl.files.File'> from /.
del combined.obs['desc_0.4']
del combined.obs['desc_0.6']
del combined.obs['desc_0.2']
del combined.obs['desc_1.4']
del combined.obs['desc_2.0']

combined.write_h5ad("combine_strctd.h5ad")

combined=sc.read_h5ad("combine_strctd.h5ad")
combined

print(combined.X)   #not norm rawdata

combined.var_names

combined.shape

## 分开traindata/validation

import random

traindata = anndata.AnnData()
# 从原始anndata对象中获取观测点的索引:train:validation = 6:4
random_indices = random.sample(range(combined.shape[0]), int(combined.shape[0]*0.6))
# 根据索引抽取对应的观测数据
traindata = combined[random_indices, :]
traindata

traindata.obs['batch'].value_counts()

combined.obs['batch'].value_counts()

traindata.write_h5ad("traindata.h5ad")

validation = anndata.AnnData()
# 从原始anndata对象中获取观测点的索引:train:validation = 6:4
random_indices = random.sample(range(combined.shape[0]), int(combined.shape[0]*0.4))
# 根据索引抽取对应的观测数据
validation = combined[random_indices, :]
validation

validation.write_h5ad("validation.h5ad")

validation.obs['batch'].value_counts()

##下采样

#平均表达个数=sc/st

sc_combine.obs['nFeature_RNA']

st_combine.obs['nFeature_Spatial']

scaverages=sc_combine.obs['nFeature_RNA'].mean()
print(scaverages)
staverages=st_combine.obs['nFeature_Spatial'].mean()   
print(staverages)
alpha=scaverages/staverages 
alpha 
#231.72607584987227
#1121.7586926286508
#0.20657390700210365

###下采样####################

def lower_sample_data(rawdata, alpha):
    """
    将矩阵中的每个元素有概率 p 保留原值，有概率 1-p 变成 0
    """
    #创建一个与原始矩阵大小相同的随机矩阵
    random_matrix=np.random.choice([0,1], size=(rawdata.shape[0],rawdata.shape[1]), replace=True, p=[1-alpha,alpha])
    # 将原始矩阵中的元素与随机矩阵中的元素按位相乘
    result = rawdata * random_matrix
    return result


validation=sc.read_h5ad("validation.h5ad")
traindata=sc.read_h5ad("traindata.h5ad")

traindata

traindatasubset_sc=traindata[traindata.obs['batch']=='sc']
traindatasubset_st=traindata[traindata.obs['batch']=='st']
traindatasubset_st

lower_sample_data_1=lower_sample_data(pd.DataFrame(traindatasubset_st.X.todense()), alpha)
lower_sample_data_1
lower_sample_data_1.index=traindatasubset_st.obs_names
lower_sample_data_1.index
lower_sample_data_1.columns=traindatasubset_st.var_names
lower_sample_data_1.columns
traindata_st=sc.AnnData(lower_sample_data_1)
traindata_st
traindata_st.obs["batch"]="st"

int(traindatasubset_sc.shape[0]/traindatasubset_st.shape[0])

###循环抽样sc/st次
names = locals()
for i in range(1,int(traindatasubset_sc.shape[0]/traindatasubset_st.shape[0])+1):
    print(i)
    names['lower_sample_data_%s'%i]=lower_sample_data(pd.DataFrame(traindatasubset_st.X.todense()), alpha)
    names['lower_sample_data_%s'%i].index=traindatasubset_st.obs_names
    names['lower_sample_data_%s'%i].columns=traindatasubset_st.var_names
    names['traindata_st_%s'%i]=sc.AnnData(lower_sample_data_1)
    names['traindata_st_%s'%i].obs["batch"]="st"
    names['traindata_st_%s'%i].obs["desc_0.8"]=traindatasubset_st.obs['desc_0.8']
    names['traindata_st_%s'%i].obs["batch_sub"]='st_%s'%i
    

##
traindata_st_65

#sc_combine.X=sc_combine.X.toarray()

import anndata as ad
names = locals()
traindata_combine_1=traindata_st_1
for i in range(1,int(traindatasubset_sc.shape[0]/traindatasubset_st.shape[0])):
    print(i)
    names['traindata_combine_%s'%(i+1)]=ad.concat([names['traindata_combine_%s'%i],names['traindata_st_%s'%(i+1)]], join="outer")


traindata_combine_65

traindata_combine_65.write_h5ad("traindata_undersample_st.h5ad")

sc_combine

#traindata_combine_26=sc.read_h5ad("traindata_undersample_st.h5ad")
traindata_combine=ad.concat([sc_combine,traindata_combine_65],join="outer")
traindata_combine

traindata_combine.obs

traindata_combine.X #需要转换稀疏阵为array，否则desc报错无法转换为tensor


#保存报错TypeError: Can't implicitly convert non-string objects to strings  Above error raised while writing key 'desc_0.4' of <class 'h5py._hl.group.Group'> from /.  Above error raised while writing key 'obs' of <class 'h5py._hl.files.File'> from /.
del traindata_combine.obs['desc_0.4']
del traindata_combine.obs['desc_0.6']
del traindata_combine.obs['desc_0.2']
del traindata_combine.obs['desc_1.4']
del traindata_combine.obs['desc_2.0']


traindata_combine.write_h5ad("traindata_undersample.h5ad")

##validation

##validation
#desc_0.8
validationsubset_sc=validation[validation.obs['batch']=='sc']
validationsubset_st=validation[validation.obs['batch']=='st']
print(validationsubset_st)
lower_sample_data_1=lower_sample_data(pd.DataFrame(validationsubset_st.X.todense()), alpha)
lower_sample_data_1
lower_sample_data_1.index=validationsubset_st.obs_names
lower_sample_data_1.index
lower_sample_data_1.columns=validationsubset_st.var_names
lower_sample_data_1.columns
validation_st=sc.AnnData(lower_sample_data_1)
validation_st
validation_st.obs["batch"]="st"
###循环抽样sc/st次
names = locals()
for i in range(1,int(validationsubset_sc.shape[0]/validationsubset_st.shape[0])+1):
    print(i)
    names['lower_sample_data_%s'%i]=lower_sample_data(pd.DataFrame(validationsubset_st.X.todense()), alpha)
    names['lower_sample_data_%s'%i].index=validationsubset_st.obs_names
    names['lower_sample_data_%s'%i].columns=validationsubset_st.var_names
    names['validation_st_%s'%i]=sc.AnnData(lower_sample_data_1)
    names['validation_st_%s'%i].obs["batch"]="st"
    names['validation_st_%s'%i].obs["desc_0.8"]=validationsubset_st.obs['desc_0.8']
    names['validation_st_%s'%i].obs["batch_sub"]='st_%s'%i

import anndata as ad
names = locals()
validation_combine_1=validation_st_1
for i in range(1,int(validationsubset_sc.shape[0]/validationsubset_st.shape[0])):
    print(i)
    names['validation_combine_%s'%(i+1)]=ad.concat([names['validation_combine_%s'%i],names['validation_st_%s'%(i+1)]], join="outer")

a=int(validationsubset_sc.shape[0]/validationsubset_st.shape[0])
a

validation_combine_65

validation_combine_65.write_h5ad("validation_undersample_st.h5ad")

validation_combine_65=sc.read_h5ad("validation_undersample_st.h5ad")
validation_combine=ad.concat([sc_combine,validation_combine_65],join="outer")
validation_combine

del validation_combine.obs['desc_0.4']
del validation_combine.obs['desc_0.6']
del validation_combine.obs['desc_0.2']
del validation_combine.obs['desc_1.4']
del validation_combine.obs['desc_2.0']


validation_combine.write_h5ad("validation_undersample.h5ad")

###########################################################

#随机数列
percent=np.random.choice([0,1], size=(st.shape[1]), replace=True, p=[0.4212802,0.5787198])
percent

percent.shape

st.var.index

type(st.X)

#对每个spot，按概率随机表达一定个数的genes

def lower_sample_data(rowdata, alpha):
    """
    将矩阵中的每个元素有概率 p 保留原值，有概率 1-p 变成 0
    """
    #创建一个与原始矩阵大小相同的随机矩阵
    random_matrix=np.random.choice([0,1], size=(st.shape[0],st.shape[1]), replace=True, p=[1-alpha,alpha])
    # 将原始矩阵中的元素与随机矩阵中的元素按位相乘
    result = rowdata * random_matrix
    return result


st.X.todense()

lower_sample_data(pd.DataFrame(st.X.todense()), 0.5)

#############################################################################

import random
def lower_sample_data(rowdata, alpha):
    newX=pd.DataFrame()
    #newX=csr_matrix
    #for index, row in rowdata.iterrows():
    #for x in range(len(row)):
    for x in rowdata:
            if random.random() < alpha:
                print(x)
                newX.append(x)
            else:
                newX=0
        return newX
